Handle widget update properly
Consider the following code in a form:
```
DateTimePickerFormField(
  enabled: _isEnabled, 
  // ....
);
```
Where the `_isEnabled` parameter is changeable by another field of the form:
```
CheckboxListTile(
  // ...
  onChanged: (value) {
    setState(() {
      _isEnabled = value;
    });
  },
);
```

When the checkbox is tapped, `setState()` is called and the widget tree is rebuilt. However, currently the state of `DateTimePickerFormField` won't change because the `State` object of the `DateTimePickerTextFormField` is bound to the original widget that creates it, so the new `DateTimePickerFormField` widget (with an updated `enabled` value) is not used at all.

This pull request fixes this by
 * removes all references to `parent` (the original widget). `widget` is used instead.
 * implements `didUpdateWidget()` to connect the event listeners to the new widget (and disconnect from the old one).
 * takes care of the case when the datetime format is changed in the new widget and update the text field content accordingly in `didUpdateWidget()`.

I didn't check every function which may need to be updated with the new widget like the `format` param, but this is at least more correct than the current behavior.