Removing render overflow error when keyboard open
When another text input is open in the form, and the keyboard is open, and then the user selects the date picker, then a Render Flex overflow happens. I think it is related to this issue:  https://github.com/flutter/flutter/issues/18672

If we instead try to dismiss the keyboard and then show the datepicker, the overflow error doesn't happen.

Also, what about removing the print statements?